config/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.27
djangorestframework==3.15.2
exceptiongroup==1.3.1
factory-boy==3.3.3
Faker==35.2.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.4
tomli==2.3.0
typing-extensions==4.13.2
